2020.07.30 21:00 Code Operation Debug #20
TODO :
1) Init FIFO - Done
2) VLCGroup() - Done
3) UpdateHistoryElement - Done
4) addbits, getbits Problem - Done
5) Flatness Test Part Out of Bound Error - Done
6) Final Line Processing - Done
7) FIFO Overflow Problem - Done
8) SamplePredict indexing Problem - Done
9) Syntax Edited - Done
10) Shifter Structure Changed - Done
11) When is Shifter Fullness increased ?? - Done
12) AttributeError: 'numpy.int32' object has no attribute 'bit_length' - Done
13) DSC Write Index Problem - Done
14) Pixel Value Debug ... Priority!!!
15) Output Buffer Write Problem... ONLY WRITES ZERO VALUE!! - Done
16) stQp, 'd' in putbit Value Type Error - Done
17) Line Buffer Value Initialize - Done
18) Zero Padding Position Problems
19) RateControl Fullness Problem

+++ PPS_readnwrite.py
@@ -98,7 +98,7 @@ def parse_pps(path = "w1.dsc", PRINT_PPS_OPT = False):
         ])
 
         # rc_range_parameters, (15, 3) shape ndarray
-        rc_range_parameters = np.zeros(shape=(15, 3))
+        rc_range_parameters = np.zeros(shape = (15, 3), dtype = np.int32)
         for i in range(rc_range_parameters.shape[0]):
             tmp = int.from_bytes(f.read(2), byteorder = 'big', signed = 0)
             rc_range_parameters[i, :] = [tmp >> 11, (tmp >> 6) & 0b11111, tmp & 0b111111]
